made portions of the graph change color according to their value
If the mood is >=15, the portion of the graph will be green, 7>=mood>15 orange, else red.
I had to compute and add dummy points to the graph to make the line change color when it crosses the threshold values. Have to add 2 dummy points when it jumps from bad to good without any point in medium.